health-report-analyzer-2/requirements.txt
streamlit==1.38.0
pandas==2.2.2
google-generativeai==0.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1